WIP: Migrate uint/ByteBuf interactions
Migrate use of Uint helpers provided by yangtools, which offer
better performance and ergonomics. This is the first step, migrating
production code, so that it is validated using existing tests.

JIRA: OPNFLWPLUG-1099

+++ openflowjava/openflow-protocol-impl/src/main/java/org/opendaylight/openflowjava/protocol/impl/deserialization/match/OxmVlanVidDeserializer.java
@@ -18,6 +18,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.oxm.rev150225.matc
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.oxm.rev150225.match.entries.grouping.MatchEntryBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.oxm.rev150225.match.entry.value.grouping.match.entry.value.VlanVidCaseBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.oxm.rev150225.match.entry.value.grouping.match.entry.value.vlan.vid._case.VlanVidBuilder;
+import org.opendaylight.yangtools.yang.common.Uint16;
 
 /**
  * Translates OxmVlanVid messages.
@@ -39,7 +40,7 @@ public class OxmVlanVidDeserializer extends AbstractOxmMatchEntryDeserializer
         VlanVidBuilder vlanBuilder = new VlanVidBuilder();
         int vidEntryValue = input.readUnsignedShort();
         vlanBuilder.setCfiBit((vidEntryValue & 1 << 12) != 0); // cfi is 13-th bit
-        vlanBuilder.setVlanVid(vidEntryValue & (1 << 12) - 1); // value without 13-th bit
+        vlanBuilder.setVlanVid(Uint16.valueOf(vidEntryValue & (1 << 12) - 1)); // value without 13-th bit
         if (builder.isHasMask()) {
             vlanBuilder.setMask(OxmDeserializerHelper
                     .convertMask(input, EncodeConstants.SIZE_OF_SHORT_IN_BYTES));
